Mod: Add new features and cleanup code
Cleans up the overall code structure of the mod and adds full support
for pretty much playing the entire game in VR.

Joystick directions select buttons marked 1, 2, or 3. A joystick click
triggers any "oops" button (don't want to end anything if you move the
joystick the wrong way)

The top face button acts as a pause button and sends the safeword command

The bottom face button hides and shows the UI depending on current UI state.

On stakes/choice UI, choices can be selected by joystick movement and
confirmed via clicking the joystick.

Also fix the bug where the mod won't activate after a session is finished.

More cleanup is needed, but this is a start.